requirements.txt
numpy==1.23.4
pandas==1.2.4
Pillow==9.0.1
timm==0.9.8
torch==1.11.0
torchvision==0.12.0
